The init.sh script supports libxulsword-only, and no-windows settings

+++ init.sh
@@ -28,8 +28,10 @@ PKG_DEPS="$PKG_DEPS debhelper libboost-dev"
 # for VM build
 if [ "$CONTEXT" = "guest" ]; then PKG_DEPS="$PKG_DEPS libxshmfence1 libglu1 libnss3-dev libgdk-pixbuf2.0-dev libgtk-3-dev libxss-dev libasound2"; fi
 
-# BUILD DEPENDENCIES (for cross compiling libxulsword as a Windows dll)
-PKG_DEPS="$PKG_DEPS mingw-w64 mingw-w64-tools wine wine32-preloader"
+if [ "$WINMACHINE" != "no" ]; then
+  # BUILD DEPENDENCIES (for cross compiling libxulsword as a Windows dll)
+  PKG_DEPS="$PKG_DEPS mingw-w64 mingw-w64-tools wine wine32-preloader"
+fi
 
 # BUILD DEPENDENCIES (for cross compiling libxulsword as MacOS dylib)
 # PKG_DEPS="$PKG_DEPS llvm-12 llvm-12-linker-tools llvm-12-tools clang-12 lldb-12 lld-12"
@@ -67,25 +69,31 @@ if [ "$CONTEXT" = "guest" ]; then export XULSWORD="$HOME/src/xulsword"; else exp
 cd "$XULSWORD"
 export CPP="$XULSWORD/Cpp"
 
-# Install node.js using nvm so our dev environment can use the latest
-# LTS version of node.js. Then install yarn and dependant node modules.
-export NVM_DIR="$HOME/.nvm"
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.38.0/install.sh | bash
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
-[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
+if [ "$LIBXULSWORD_ONLY" = "no" ]; then
+  # Install node.js using nvm so our dev environment can use the latest
+  # LTS version of node.js. Then install yarn and dependant node modules.
+  export NVM_DIR="$HOME/.nvm"
+  curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.38.0/install.sh | bash
+  [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
+  [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
 
-nvm install 18.12.1
-nvm use 18.12.1
-npm i --global yarn
+  nvm install 18.12.1
+  nvm use 18.12.1
+  npm i --global yarn
+fi
 
 # Create a local Cpp installation directory where compiled libraries will
 # be installed for libxulsword linking.
 if [ ! -e "$XULSWORD/Cpp/install" ]; then mkdir "$XULSWORD/Cpp/install"; fi
-if [ ! -e "$XULSWORD/Cpp/install.$XCWD" ]; then mkdir "$XULSWORD/Cpp/install.$XCWD"; fi
+if [ "$WINMACHINE" != "no" ]; then
+  if [ ! -e "$XULSWORD/Cpp/install.$XCWD" ]; then mkdir "$XULSWORD/Cpp/install.$XCWD"; fi
+fi
 
 # Create a local lib directory where libxulsword will be installed
 if [ ! -e "$XULSWORD/Cpp/lib" ]; then mkdir "$XULSWORD/Cpp/lib"; fi
-if [ ! -e "$XULSWORD/Cpp/lib.$XCWD" ]; then mkdir "$XULSWORD/Cpp/lib.$XCWD"; fi
+if [ "$WINMACHINE" != "no" ]; then
+  if [ ! -e "$XULSWORD/Cpp/lib.$XCWD" ]; then mkdir "$XULSWORD/Cpp/lib.$XCWD"; fi
+fi
 
 # Create an archive directory to cache source code
 ARCHIVEDIR="$XULSWORD/archive"
@@ -140,29 +148,33 @@ if [ ! -e "$CPP/$dirout" ]; then
   ln -s ./build/zconf.h ../zconf.h
 fi
 # CROSS COMPILE ZLIB TO WINDOWS
-dirout="zlib.$XCWD"
-if [ ! -e "$CPP/$dirout" ]; then
-  getSource
-  cmake $DBG -G "Unix Makefiles" -D CMAKE_TOOLCHAIN_FILE="$CPP/windows/toolchain.cmake" ..
-  make DESTDIR="$CPP/install.$XCWD" install
+if [ "$WINMACHINE" != "no" ]; then
+  dirout="zlib.$XCWD"
+  if [ ! -e "$CPP/$dirout" ]; then
+    getSource
+    cmake $DBG -G "Unix Makefiles" -D CMAKE_TOOLCHAIN_FILE="$CPP/windows/toolchain.cmake" ..
+    make DESTDIR="$CPP/install.$XCWD" install
+  fi
 fi
 ########################################################################
 
 
 ########################################################################
 # CROSS-COMPILE BOOST TO WINDOWS FOR CLUCENE
-url=
-gzfile="boost_1_80_0.tar.gz"
-dirin="boost_1_80_0"
-dirout="boost-${TOOLCHAIN_PREFIX}-${XCWD}"
-BOOSTDIR="$CPP/boost-${TOOLCHAIN_PREFIX}-${XCWD}"
-if [ ! -e "$BOOSTDIR" ]; then
-  getSource
-  # CROSS COMPILE TO WINDOWS 64 BIT:
-  cd ..
-  echo "using gcc :  : ${TOOLCHAIN_PREFIX}-g++ ;" > user-config.jam
-  ./bootstrap.sh
-  ./b2 --user-config=./user-config.jam --prefix=./$XCWD target-os=windows address-model=$ADDRESS_MODEL variant=release install
+if [ "$WINMACHINE" != "no" ]; then
+  url=
+  gzfile="boost_1_80_0.tar.gz"
+  dirin="boost_1_80_0"
+  dirout="boost-${TOOLCHAIN_PREFIX}-${XCWD}"
+  BOOSTDIR="$CPP/boost-${TOOLCHAIN_PREFIX}-${XCWD}"
+  if [ ! -e "$BOOSTDIR" ]; then
+    getSource
+    # CROSS COMPILE TO WINDOWS 64 BIT:
+    cd ..
+    echo "using gcc :  : ${TOOLCHAIN_PREFIX}-g++ ;" > user-config.jam
+    ./bootstrap.sh
+    ./b2 --user-config=./user-config.jam --prefix=./$XCWD target-os=windows address-model=$ADDRESS_MODEL variant=release install
+  fi
 fi
 ########################################################################
 
@@ -182,17 +194,19 @@ if [ ! -e "$CPP/$dirout" ]; then
   make DESTDIR="$CPP/install" install
 fi
 # CROSS COMPILE LIBCLUCENE TO WINDOWS
-dirout="clucene.$XCWD"
-if [ ! -e "$CPP/$dirout" ]; then
-  getSource
-  # Stop this dumb clucene error for searches beginning with a wildcard, which results in a core dump.
-  sed -i 's/!allowLeadingWildcard/!true/g' "$CPP/$dirout/src/core/CLucene/queryParser/QueryParser.cpp"
-  cd "$CPP"
-  patch -s -p0 -d "$CPP/$dirout" < "$CPP/windows/clucene-src.patch"
-  cd "$CPP/$dirout/build"
-  cmake $DBG -DCMAKE_TOOLCHAIN_FILE="$CPP/windows/toolchain.cmake" -C "$CPP/windows/clucene-TryRunResult-${GCCSTD}.cmake" -D CMAKE_USE_PTHREADS_INIT=OFF -D BUILD_STATIC_LIBRARIES=ON -D ZLIB_INCLUDE_DIR="$CPP/install.$XCWD/usr/local/include" -D Boost_INCLUDE_DIR="$BOOSTDIR/$XCWD/include" -D ZLIB_LIBRARY="$CPP/install.$XCWD/usr/local/lib/libzlibstatic.a" ..
-  patch -s -p0 -d "$CPP/$dirout" < "$CPP/windows/clucene-build.patch"
-  make DESTDIR="$CPP/install.$XCWD" install
+if [ "$WINMACHINE" != "no" ]; then
+  dirout="clucene.$XCWD"
+  if [ ! -e "$CPP/$dirout" ]; then
+    getSource
+    # Stop this dumb clucene error for searches beginning with a wildcard, which results in a core dump.
+    sed -i 's/!allowLeadingWildcard/!true/g' "$CPP/$dirout/src/core/CLucene/queryParser/QueryParser.cpp"
+    cd "$CPP"
+    patch -s -p0 -d "$CPP/$dirout" < "$CPP/windows/clucene-src.patch"
+    cd "$CPP/$dirout/build"
+    cmake $DBG -DCMAKE_TOOLCHAIN_FILE="$CPP/windows/toolchain.cmake" -C "$CPP/windows/clucene-TryRunResult-${GCCSTD}.cmake" -D CMAKE_USE_PTHREADS_INIT=OFF -D BUILD_STATIC_LIBRARIES=ON -D ZLIB_INCLUDE_DIR="$CPP/install.$XCWD/usr/local/include" -D Boost_INCLUDE_DIR="$BOOSTDIR/$XCWD/include" -D ZLIB_LIBRARY="$CPP/install.$XCWD/usr/local/lib/libzlibstatic.a" ..
+    patch -s -p0 -d "$CPP/$dirout" < "$CPP/windows/clucene-build.patch"
+    make DESTDIR="$CPP/install.$XCWD" install
+  fi
 fi
 ########################################################################
 
@@ -213,16 +227,18 @@ if [ ! -e "$CPP/$dirout" ]; then
   make DESTDIR="$CPP/install" install
 fi
 # CROSS COMPILE LIBSWORD TO WINDOWS
-dirout="sword.$XCWD"
-if [ ! -e "$CPP/$dirout" ]; then
-  getSource
-  # SWORD's CMakeLists.txt requires clucene-config.h be located in a weird directory:
-  cp -r "$CPP/install.$XCWD/usr/local/include/CLucene" "$CPP/install.$XCWD/usr/local/lib"
-  cd "$CPP"
-  patch -s -p0 -d "$CPP/$dirout" < "$CPP/windows/libsword-src.patch"
-  cd "$CPP/$dirout/build"
-  cmake $DBG -DCMAKE_TOOLCHAIN_FILE="$CPP/windows/toolchain.cmake" -D SWORD_NO_ICU="No" -D LIBSWORD_LIBRARY_TYPE="Static" -D CLUCENE_LIBRARY="$CPP/install.$XCWD/usr/local/lib/libclucene-core.dll.a" -D ZLIB_LIBRARY="$CPP/install.$XCWD/usr/local/lib/libzlibstatic.a" -D CLUCENE_LIBRARY_DIR="$CPP/install.$XCWD/usr/local/include" -D CLUCENE_INCLUDE_DIR="$CPP/install.$XCWD/usr/local/include" -D ZLIB_INCLUDE_DIR="$CPP/install.$XCWD/usr/local/include" -DSWORD_BUILD_UTILS="No" ..
-  make DESTDIR="$CPP/install.$XCWD" install
+if [ "$WINMACHINE" != "no" ]; then
+  dirout="sword.$XCWD"
+  if [ ! -e "$CPP/$dirout" ]; then
+    getSource
+    # SWORD's CMakeLists.txt requires clucene-config.h be located in a weird directory:
+    cp -r "$CPP/install.$XCWD/usr/local/include/CLucene" "$CPP/install.$XCWD/usr/local/lib"
+    cd "$CPP"
+    patch -s -p0 -d "$CPP/$dirout" < "$CPP/windows/libsword-src.patch"
+    cd "$CPP/$dirout/build"
+    cmake $DBG -DCMAKE_TOOLCHAIN_FILE="$CPP/windows/toolchain.cmake" -D SWORD_NO_ICU="No" -D LIBSWORD_LIBRARY_TYPE="Static" -D CLUCENE_LIBRARY="$CPP/install.$XCWD/usr/local/lib/libclucene-core.dll.a" -D ZLIB_LIBRARY="$CPP/install.$XCWD/usr/local/lib/libzlibstatic.a" -D CLUCENE_LIBRARY_DIR="$CPP/install.$XCWD/usr/local/include" -D CLUCENE_INCLUDE_DIR="$CPP/install.$XCWD/usr/local/include" -D ZLIB_INCLUDE_DIR="$CPP/install.$XCWD/usr/local/include" -DSWORD_BUILD_UTILS="No" ..
+    make DESTDIR="$CPP/install.$XCWD" install
+  fi
 fi
 ########################################################################
 
@@ -235,7 +251,7 @@ if [ ! -e "$CPP/build" ]; then
   cmake $DBG -D SWORD_NO_ICU="No" -D CMAKE_INCLUDE_PATH="$CPP/install/usr/local/include" -D CMAKE_LIBRARY_PATH="$CPP/install/usr/local/lib" ..
   make DESTDIR="$CPP/install" install
 
-  # Install the DLL and all ming dependencies (those beyond the node executable) and strip them
+  # Install the lib and all dependencies and strip them
   LIBDIR="$CPP/lib"
   if [ -e "$LIBDIR" ]; then rm -rf "$LIBDIR"; fi
   mkdir "$LIBDIR"
@@ -254,39 +270,41 @@ if [ ! -e "$CPP/build" ]; then
   fi
 fi
 # CROSS COMPILE LIBXULSWORD TO WINDOWS
-if [ ! -e "$CPP/build.$XCWD" ]; then
-  mkdir "$CPP/build.$XCWD"
-  cd "$CPP/build.$XCWD"
-  cmake $DBG -DCMAKE_TOOLCHAIN_FILE="$CPP/windows/toolchain.cmake" -D SWORD_NO_ICU="No" -D CMAKE_INCLUDE_PATH="$CPP/install.$XCWD/usr/local/include" -D CMAKE_LIBRARY_PATH="$CPP/install.$XCWD/usr/local/lib" ..
-  make DESTDIR="$CPP/install.$XCWD" install
+if [ "$WINMACHINE" != "no" ]; then
+  if [ ! -e "$CPP/build.$XCWD" ]; then
+    mkdir "$CPP/build.$XCWD"
+    cd "$CPP/build.$XCWD"
+    cmake $DBG -DCMAKE_TOOLCHAIN_FILE="$CPP/windows/toolchain.cmake" -D SWORD_NO_ICU="No" -D CMAKE_INCLUDE_PATH="$CPP/install.$XCWD/usr/local/include" -D CMAKE_LIBRARY_PATH="$CPP/install.$XCWD/usr/local/lib" ..
+    make DESTDIR="$CPP/install.$XCWD" install
 
-  # Install the DLL and all ming dependencies beyond the node executable and strip them
-  LIBDIR="$CPP/lib.$XCWD"
-  if [ -e "$LIBDIR" ]; then rm -rf "$LIBDIR"; fi
-  GCCDLL=libgcc_s_seh-1.dll
-  if [[ "$XCWD" == "32win" ]]; then
-    GCCDLL=libgcc_s_sjlj-1.dll
-  fi
-  mkdir "$LIBDIR"
-  cp "$CPP/install.$XCWD/usr/local/bin/libxulsword-static.dll" "$LIBDIR"
-  cp "/usr/lib/gcc/${TOOLCHAIN_PREFIX}/9.3-${GCCSTD}/$GCCDLL" "$LIBDIR"
-  cp "/usr/lib/gcc/${TOOLCHAIN_PREFIX}/9.3-${GCCSTD}/libstdc++-6.dll" "$LIBDIR"
-  cp "/usr/${TOOLCHAIN_PREFIX}/lib/libwinpthread-1.dll" "$LIBDIR"
-  gendef - "$CPP/install.$XCWD/usr/local/bin/libxulsword-static.dll" > "$LIBXULSWORD/lib/libxulsword.def"
-  if [ -z "$DBG" ]; then
-    strip "$LIBDIR/"*
-  fi
-  chmod ugo+x "$LIBDIR/"*
+    # Install the DLL and all ming dependencies beyond the node executable and strip them
+    LIBDIR="$CPP/lib.$XCWD"
+    if [ -e "$LIBDIR" ]; then rm -rf "$LIBDIR"; fi
+    GCCDLL=libgcc_s_seh-1.dll
+    if [[ "$XCWD" == "32win" ]]; then
+      GCCDLL=libgcc_s_sjlj-1.dll
+    fi
+    mkdir "$LIBDIR"
+    cp "$CPP/install.$XCWD/usr/local/bin/libxulsword-static.dll" "$LIBDIR"
+    cp "/usr/lib/gcc/${TOOLCHAIN_PREFIX}/9.3-${GCCSTD}/$GCCDLL" "$LIBDIR"
+    cp "/usr/lib/gcc/${TOOLCHAIN_PREFIX}/9.3-${GCCSTD}/libstdc++-6.dll" "$LIBDIR"
+    cp "/usr/${TOOLCHAIN_PREFIX}/lib/libwinpthread-1.dll" "$LIBDIR"
+    gendef - "$CPP/install.$XCWD/usr/local/bin/libxulsword-static.dll" > "$LIBXULSWORD/lib/libxulsword.def"
+    if [ -z "$DBG" ]; then
+      strip "$LIBDIR/"*
+    fi
+    chmod ugo+x "$LIBDIR/"*
 
-  # If this is Vagrant, then copy the finished library to the host machine
-  if [ "$CONTEXT" = "guest" ]; then
-    HLIBDIR="/vagrant/Cpp/lib.$XCWD"
-    if [ -e "$HLIBDIR" ]; then rm -rf "$HLIBDIR"; fi
-    cp -r "$LIBDIR" "$HLIBDIR"
+    # If this is Vagrant, then copy the finished library to the host machine
+    if [ "$CONTEXT" = "guest" ]; then
+      HLIBDIR="/vagrant/Cpp/lib.$XCWD"
+      if [ -e "$HLIBDIR" ]; then rm -rf "$HLIBDIR"; fi
+      cp -r "$LIBDIR" "$HLIBDIR"
+    fi
   fi
 fi
 ########################################################################
-
+if [ "$LIBXULSWORD_ONLY" = "yes" ]; then exit 0; fi
 
 ########################################################################
 # WRAP UP
